refactor(calendar): remove src/calendar.rs after modularization and fix for Rust module conventions
